Import redir v2.2
Added support for redirecting PORT mode ftp connections in addition
to PASV mode ftp redirection.  Thus --ftp is now split into

    --ftp={port,pasv,both}

to determine what should be redirected.  The original version of this
patch was submitted by Harald Holzer <harald.holzer@eunet.at>.

Also added, the --connect option, which is useful if you're bouncing
your connections through an HTTP proxy server.  Use as

    --connect=host:port

and this will be the CONNECT line sent to the proxy.

Signed-off-by: Sam Creasey <sammy@oh.verio.com>