Import redir v2.2.1
Fix a bug in do_accept() where non-fatal error codes returned by
accept() would cause redir to terminate entirely.  I had recieved
reports of this behavior but was unable to find it until sammy.net had
to handle the load of the redir 2.2 update using redir. :)
All non-fatal error codes might not be covered.  But it "got better".

Integrate a patch by Emmanuel Chantréau <echant@maretmanu.org> which
provides traffic shaping functionality to redir.  Interesting stuff.
I've not tested this in detail personally.

Also, added the ability to compile redir with lesser functionality for
speed.  This is documented in the README.

Signed-off-by: Sam Creasey <sammy@oh.verio.com>