Do gethostbyname() before every connect(), names may change at runtime
This patch is a bit messy, increasing the amount of code duplication.
The overall goal is however to call gethostbyname() in each new session
started in regular (non-inetd) mode, because DNS may change over the
life space of this process.

A later patch will address the code duplication beteen inetd and regular
server mode.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>