Log to stderr when running in foreground w/o syslog
On some UNIX systems, in particular GNU, openlog() supports LOG_PERROR
which makes syslog() print to stderr as well as the syslog.  This can
be handy for some users, and use-cases such as debugging.

This patch enables LOG_PERROR when running in foreground without the
specific syslog command line option.  So, we can now call that option
"silent" as well.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>