Add --enable-compat for v2.x command line support
This patch is an attempt to ease the transition of redir3 into Debian.
The only package that depends on redir is vagrant-lxc, which obviously
uses the old command line syntax.

The `--enable-compat` configure option enables the old syntax required
by vagrant-lxc, nothing more.  Using any of those command line options
also triggers redir to run in the foreground and disable syslog.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>