Fix wait features to use microseconds
The wait features (`--random-wait` and `--max-bandwidth`) are based on
the assumption that the 'struct timeval' parameter in 'select' uses
milliseconds for its 'tv_usec' member, but it actually uses
microseconds. This has a dramatic impact on the accuracy of individual
delays below one second - they are too short by a factor of 1000.